Time: 56 ms (91.32%) | Memory: 195.5 MB (49.21%) - LeetSync

+++ 3886-count-number-of-trapezoids-i/count-number-of-trapezoids-i.cpp
@@ -3,7 +3,7 @@ public:
     static const long long MOD = 1e9 + 7;
 
     int countTrapezoids(vector<vector<int>>& points) {
-        unordered_map<long long, long long> countY;
+        unordered_map<int, long long> countY;
 
         // Count points at each y-level
         for (auto &p : points) {
